Branding initial changes

+++ idas.chox.web/src/main/java/idas/chox/web/actions/InsurerAction.java
@@ -151,7 +151,7 @@ public class InsurerAction extends BaseAction implements ModelDriven<Insurer>, P
 
     @Secured({"ROLE_CHOX_ADMIN", "ROLE_INS_ADMIN"})
     public String updateInsurer() {
-        if (getIsInsurer() && model.getId().intValue() != getAuthenticatedUser().getInsurer().getId().intValue()) {
+        if (getIsInsurer() && model.getId() != getAuthenticatedUser().getInsurer().getId().intValue()) {
             throw new AccessDeniedException("Cannot update other Insurer");
         }
 
@@ -226,6 +226,4 @@ public class InsurerAction extends BaseAction implements ModelDriven<Insurer>, P
     public void setOriginalName(String originalName) {
         this.originalName = originalName;
     }
-    
-    
 }
